+++ src/Libraries Projects/Milochau.Core.AspNetCore/Infrastructure/Middlewares/SystemApplicationMiddleware.cs
@@ -2,7 +2,6 @@
 using Milochau.Core.Abstractions;
 using Milochau.Core.Infrastructure.Features.Application;
 using System;
-using System.Globalization;
 using System.Reflection;
 using System.Threading.Tasks;
 
@@ -32,47 +31,17 @@ namespace Milochau.Core.AspNetCore.Infrastructure.Middlewares
             };
         }
 
-        private Task AssemblyInformationAsync(HttpContext httpContext)
+        private static Task AssemblyInformationAsync(HttpContext httpContext)
         {
             var assembly = Assembly.GetEntryAssembly();
-            var response = new AssemblyResponse
-            {
-                Company = assembly.GetCustomAttribute<AssemblyCompanyAttribute>()?.Company,
-                Configuration = assembly.GetCustomAttribute<AssemblyConfigurationAttribute>()?.Configuration,
-                Copyright = assembly.GetCustomAttribute<AssemblyCopyrightAttribute>()?.Copyright,
-                Description = assembly.GetCustomAttribute<AssemblyDescriptionAttribute>()?.Description,
-                FileVersion = assembly.GetCustomAttribute<AssemblyFileVersionAttribute>()?.Version,
-                InformationalVersion = assembly.GetCustomAttribute<AssemblyInformationalVersionAttribute>()?.InformationalVersion,
-                Product = assembly.GetCustomAttribute<AssemblyProductAttribute>()?.Product
-            };
-
-            response.IsLocal = AssemblyResponse.IsLocalRegex.IsMatch(response.InformationalVersion);
-            response.BuildId = AssemblyResponse.BuildRegex.Match(response.InformationalVersion).Groups[1].Value;
-            response.BuildSourceVersion = AssemblyResponse.BuildRegex.Match(response.InformationalVersion).Groups[2].Value;
+            var response = new AssemblyResponse(assembly);
 
             return BaseApplicationMiddleware.WriteResponseAsJsonAsync(httpContext, response);
         }
 
         private Task EnvironmentAsync(HttpContext httpContext)
         {
-            var response = new EnvironmentResponse
-            {
-                ApplicationName = applicationHostEnvironment.ApplicationName,
-                HostName = applicationHostEnvironment.HostName,
-                EnvironmentName = applicationHostEnvironment.EnvironmentName,
-
-                MachineName = Environment.MachineName,
-                ProcessorCount = Environment.ProcessorCount,
-                OSVersion = Environment.OSVersion.ToString(),
-                ClrVersion = Environment.Version.ToString(),
-                Is64BitOperatingSystem = Environment.Is64BitOperatingSystem,
-                Is64BitProcess = Environment.Is64BitProcess,
-
-                LocalTimeZone = TimeZoneInfo.Local.Id,
-                UtcTimeZone = TimeZoneInfo.Utc.Id,
-
-                CurrentCulture = CultureInfo.CurrentCulture.Name
-            };
+            var response = new EnvironmentResponse(applicationHostEnvironment);
             return BaseApplicationMiddleware.WriteResponseAsJsonAsync(httpContext, response);
         }
     }