+++ src/Libraries Projects/Milochau.Core.Functions/Functions/SystemApplicationFunctions.cs
@@ -4,8 +4,6 @@ using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.Http;
 using Milochau.Core.Abstractions;
 using Milochau.Core.Infrastructure.Features.Application;
-using System.Globalization;
-using System;
 
 namespace Milochau.Core.Functions.Functions
 {
@@ -24,24 +22,16 @@ namespace Milochau.Core.Functions.Functions
         [FunctionName("System-Application-Environment")]
         public IActionResult Environment([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "system/application/environment")] HttpRequest request)
         {
-            var response = new EnvironmentResponse
-            {
-                ApplicationName = applicationHostEnvironment.ApplicationName,
-                HostName = applicationHostEnvironment.HostName,
-                EnvironmentName = applicationHostEnvironment.EnvironmentName,
+            var response = new EnvironmentResponse(applicationHostEnvironment);
+            return new OkObjectResult(response);
+        }
 
-                MachineName = System.Environment.MachineName,
-                ProcessorCount = System.Environment.ProcessorCount,
-                OSVersion = System.Environment.OSVersion.ToString(),
-                ClrVersion = System.Environment.Version.ToString(),
-                Is64BitOperatingSystem = System.Environment.Is64BitOperatingSystem,
-                Is64BitProcess = System.Environment.Is64BitProcess,
-
-                LocalTimeZone = TimeZoneInfo.Local.Id,
-                UtcTimeZone = TimeZoneInfo.Utc.Id,
-
-                CurrentCulture = CultureInfo.CurrentCulture.Name
-            };
+        /// <summary>Get application asembly</summary>
+        [FunctionName("System-Application-Assembly")]
+        public IActionResult Assembly([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "system/application/assembly")] HttpRequest request)
+        {
+            var assembly = System.Reflection.Assembly.GetExecutingAssembly();
+            var response = new AssemblyResponse(assembly);
             return new OkObjectResult(response);
         }
     }