+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Add configuration file to test Functions reference file locally
 - Remove indirect references to EntityFramework Core
 - Creates a new Milochau.Core.Console library, to expose all Milochau.Core features to console applications (**BETA**)
+- Support an `OrganizationName` property in host options
 
 ## Bug fixes